refactor: 컴포넌트별 스타일 파일 분리 및 주석 제거
