+++ src/components/home/LoginModalContainer.tsx
@@ -1,4 +1,3 @@
-// components/modal/LoginModalContainer.tsx
 import React from 'react';
 import { LoginModal } from '@components/modal/LoginModal';
