+++ src/components/home/PostList/PostList.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import * as S from '@pages/home/Home.Styled';
+import * as S from './PostList.Styled';
 
 interface Post {
   id: number;
